requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-multipart
python-jose
pytz
python-pptx
langchain
langchain-google-genai
langgraph
langchain-community
pypdf
pydantic
typing-extensions
pymysql
passlib
torch
transformers
docx2txt
langchain_huggingface
pytest
